some feature

+++ app.js
@@ -1,169 +1,3 @@
-// require("dotenv").config()
-
-// const express=require("express");
-// const app=express();
-// const bcrypt=require("bcrypt")
-// const jwt=require("jsonwebtoken")
-// const cookieParser=require('cookie-parser')
-// app.use(cookieParser());
-
-// const connectToDb = require("./database/databaseConnection");
-// const Blog = require("./model/blogModel");
-
-// const {multer,storage}=require('./middleware/multerConfig');
-// const User = require("./model/userModel");
-// const isAuthenticated = require("./middleware/isAuthenticated");
-// const upload=multer({storage:storage});
-
-// app.use(express.json());
-// app.use(express.urlencoded({extended:true}));
-
-// connectToDb();
-
-// app.use(express.static("./storage"))
-// app.set('view engine','ejs')
-
-// app.get("/",async (req,res)=>{
-//   const blogs=await Blog.find()    //return array
-//   res.render("home.ejs",{blogs:blogs})
-// })
-
-// app.get("/about",(req,res)=>{
-//   const name="Sau Rav"
-//   res.render("about.ejs",{name:name})
-// })
-
-// app.get("/contact",(req,res)=>{
-//    res.render("contact.ejs")
-// })
-
-// app.get("/register",(req,res)=>{
- 
-//    res.render("register.ejs")
-// })
-
-// app.post("/register",async (req,res)=>{
-//  const {username,email,password}=req.body;
-//  await User.create({
-//   username:username,
-//   email:email,
-//   password:bcrypt.hashSync(password,12),
-//  })
-//    res.redirect("/login")
-// })
-
-
-
-// app.get("/login",(req,res)=>{
-//    res.render("login.ejs")
-// })
-
-
-// app.post("/login",async (req,res)=>{
-//   const {email,password}=req.body;
-//   const user=await User.find({email:email})
-//   if(user.length===0){
-//     res.send("Invalid email or password")
-//   }else{
-//   //  check- password
-//   const isMatched=bcrypt.compareSync(password,user[0].password)
-//   if(!isMatched){
-//     res.send("Invalid Password")
-//   }else{
-//     //require dotenv config file
-//     const token=jwt.sign({userId:user[0]._id},process.env.SECRET,{
-//       expiresIn:'20d'
-//     })
-//     res.cookie("token",token)
-//     res.send("logged in successfully")
-//   }
-//   }
-//  })
-
-
-// app.get("/blog/:id",isAuthenticated,async (req,res)=>{
-//   const id=req.params.id;
-//   const blog=await Blog.findById(id);
-//    res.render("blog.ejs",{blog:blog})
-// })
-
-
-// app.get("/deleteblog/:id",async (req,res)=>{
-//   const id=req.params.id;
-//   const remove=await Blog.findByIdAndDelete(id);
-//    res.redirect("/")
-// })
-
-
-// app.get("/editblog/:id",async (req,res)=>{
-//   const id=req.params.id;
-//   const blog=await Blog.findById(id);
-//    res.render("editblog",{blog:blog})
-// })
-
-
-// app.post("/editblog/:id",upload.single('image'),async (req,res)=>{
-//   const id=req.params.id;
-//   const fileName=req.file.filename;
-//   const {title,subtitle,description}=req.body;
-//   const blog=await Blog.findByIdAndUpdate(id,{
-//     title:title,
-//     subtitle:subtitle,
-//     description:description,
-//     image:fileName
-//   });
-//    res.redirect("/blog/"+id)
-// })
-
-
-// app.get("/search", async (req, res) => {
-//   const query = req.query.query;
-//   if (!query) {
-//       return res.redirect("/");
-//   }
-
-//   const blogs = await Blog.find({
-//       $or: [
-//           { title: new RegExp(query, 'i') },
-//           { subtitle: new RegExp(query, 'i') },
-//           { description: new RegExp(query, 'i') }
-//       ]
-//   });
-
-//   res.render("search.ejs", { blogs: blogs, query: query });
-// });
-
-
-// app.get("/createblog",isAuthenticated,(req,res)=>{
-//   res.render("createblog.ejs")
-// })
-
-
-// app.post("/createblog",upload.single('image'),async (req,res)=>{
-//   // console.log(req.body)
-//   const fileName=req.file.filename;
-  
-//   const{title,subtitle,description}=req.body;
-//   console.log(title,subtitle,description);
-//   await Blog.create({
-//     title:title,
-//     subtitle:subtitle,
-//     description:description,
-//     image:fileName
-//   })
-//   res.redirect("/")
-// })
-
-
-
-
-
-// app.listen(3000,()=>{
-//   console.log("Node Js has started at port:" + 3000);
-// })
-
-
-
 require("dotenv").config();
 
 const express = require("express");
@@ -173,8 +7,8 @@ const jwt = require("jsonwebtoken");
 const cookieParser = require("cookie-parser");
 const connectToDb = require("./database/databaseConnection");
 const Blog = require("./model/blogModel");
-const { multer, storage } = require("./middleware/multerConfig");
 const User = require("./model/userModel");
+const { multer, storage } = require("./middleware/multerConfig");
 const isAuthenticated = require("./middleware/isAuthenticated");
 const upload = multer({ storage: storage });
 
@@ -188,12 +22,15 @@ app.use(express.static("./storage"));
 app.set("view engine", "ejs");
 
 // Middleware to pass user status to views
-app.use((req, res, next) => {
+app.use(async (req, res, next) => {
     const token = req.cookies.token;
     if (token) {
         try {
             const decoded = jwt.verify(token, process.env.SECRET);
-            req.user = decoded;
+            const user = await User.findById(decoded.userId);
+            if (user) {
+                req.user = { userId: decoded.userId, username: user.username };
+            }
         } catch (err) {
             res.clearCookie("token");
         }
@@ -240,18 +77,17 @@ app.post("/login", async (req, res) => {
     if (!user) {
         return res.send("Invalid email or password");
     }
-    
+
     const isMatched = bcrypt.compareSync(password, user.password);
     if (!isMatched) {
         return res.send("Invalid Password");
     }
-    
-    const token = jwt.sign({ userId: user._id }, process.env.SECRET, {
+
+    const token = jwt.sign({ userId: user._id, username: user.username }, process.env.SECRET, {
         expiresIn: '20d',
     });
     res.cookie("token", token);
-    // res.send("logged in successfully");
-    res.redirect("/")
+    res.redirect("/");
 });
 
 app.get("/logout", (req, res) => {
@@ -265,28 +101,47 @@ app.get("/blog/:id", isAuthenticated, async (req, res) => {
     res.render("blog.ejs", { blog: blog });
 });
 
-app.get("/deleteblog/:id", async (req, res) => {
+app.get("/deleteblog/:id", isAuthenticated, async (req, res) => {
     const id = req.params.id;
+    const blog = await Blog.findById(id);
+
+    if (!blog || blog.author !== req.user.username) {
+        return res.status(403).send("You are not authorized to delete this blog.");
+    }
+
     await Blog.findByIdAndDelete(id);
     res.redirect("/");
 });
 
-app.get("/editblog/:id", async (req, res) => {
+app.get("/editblog/:id", isAuthenticated, async (req, res) => {
     const id = req.params.id;
     const blog = await Blog.findById(id);
+
+    if (!blog || blog.author !== req.user.username) {
+        return res.status(403).send("You are not authorized to edit this blog.");
+    }
+
     res.render("editblog", { blog: blog });
 });
 
-app.post("/editblog/:id", upload.single('image'), async (req, res) => {
+app.post("/editblog/:id", upload.single('image'), isAuthenticated, async (req, res) => {
     const id = req.params.id;
-    const fileName = req.file.filename;
+    const blog = await Blog.findById(id);
+
+    if (!blog || blog.author !== req.user.username) {
+        return res.status(403).send("You are not authorized to edit this blog.");
+    }
+
+    const fileName = req.file ? req.file.filename : blog.image;
     const { title, subtitle, description } = req.body;
+
     await Blog.findByIdAndUpdate(id, {
         title: title,
         subtitle: subtitle,
         description: description,
         image: fileName,
     });
+
     res.redirect("/blog/" + id);
 });
 
@@ -300,7 +155,8 @@ app.get("/search", async (req, res) => {
         $or: [
             { title: new RegExp(query, 'i') },
             { subtitle: new RegExp(query, 'i') },
-            { description: new RegExp(query, 'i') }
+            { description: new RegExp(query, 'i') },
+            { author: new RegExp(query, 'i') }
         ]
     });
 
@@ -311,7 +167,7 @@ app.get("/createblog", isAuthenticated, (req, res) => {
     res.render("createblog.ejs");
 });
 
-app.post("/createblog", upload.single('image'), async (req, res) => {
+app.post("/createblog", upload.single('image'), isAuthenticated, async (req, res) => {
     const fileName = req.file.filename;
     const { title, subtitle, description } = req.body;
     await Blog.create({
@@ -319,6 +175,7 @@ app.post("/createblog", upload.single('image'), async (req, res) => {
         subtitle: subtitle,
         description: description,
         image: fileName,
+        author: req.user.username,
     });
     res.redirect("/");
 });
